+++ CustomDropDown/Model/DropDownMode.swift
@@ -1,0 +1,14 @@
+//
+//  DropDownMode.swift
+//  CustomDropDown
+//
+//  Created by Jacqueline Alves on 22/10/20.
+//  Copyright © 2020 Amirthy Tejeshwar. All rights reserved.
+//
+
+import Foundation
+
+public enum DropDownMode: Int, CaseIterable {
+    case label
+    case imageLabel
+}
