+++ CustomDropDown/Presenter/CustomDropDownPresenter.swift
@@ -8,21 +8,6 @@
 
 import UIKit
 
-public struct DropDownConfig {
-    public var dropDownWidth: CGFloat?
-    public var dropDownTag: Int = 999
-    public var selectedLabelTag: Int = 9999
-    public var dropDownLeftRightPadding: UIEdgeInsets = UIEdgeInsets(top: 10, left: 10, bottom: 0, right: 10)
-    public var dropDownMode: DropDownMode = .label
-    
-    public init() {}
-}
-
-public enum DropDownMode: Int, CaseIterable {
-    case label
-    case imageLabel
-}
-
 public protocol CustomDropDownDataSource: class {
     func overrideDropDownView(identifier: Int) -> UIView?
     func config(identifier: Int) -> DropDownConfig
